Cleaning up in platonium files
