+++ .gitignore
@@ -81,7 +81,7 @@ PHOENIX_HiResFITS/
 WAVE_PHOENIX-ACES-AGSS-COND-2011.fits
 varsim_exomass_fitting_parameters.h5
 RRL-CEP/
-
+response_*
 
 # Compiled source
