--- python/platosim/varsim/data/meunier_19a_t1.dat
@@ -1,20 +1,0 @@
-BV SpT Teff Radius Mass
-0.49 F6 6334 1.41 1.20
-0.52 F7 6218 1.30 1.14
-0.55 F8 6112 1.22 1.09
-0.58 F9 6017 1.14 1.04
-0.60 G0 5931 1.09 1.01
-0.62 G1 5854 1.04 0.98
-0.64 G2 5784 1.00 0.95
-0.67 G3 5719 0.97 0.93
-0.69 G4 5658 0.94 0.91
-0.71 G5 5598 0.92 0.89
-0.73 G6 5536 0.89 0.87
-0.76 G7 5470 0.87 0.85
-0.79 G8 5397 0.85 0.84
-0.83 G9 5314 0.82 0.82
-0.87 K0 5215 0.80 0.80
-0.92 K1 5099 0.78 0.78
-0.99 K2 4960 0.75 0.76
-1.06 K3 4793 0.73 0.74
-1.15 K4 4594 0.71 0.72
