--- python/platosim/varsim/data/response_kepler.txt
@@ -1,31 +1,0 @@
-# Table from the Kepler Instrument Handbook:
-# https://archive.stsci.edu/files/live/sites/mast/files/home/missions-and-data/kepler/_documents/KSCI-19033-002-instrument-hb.pdf
-# lmabda, response
-# (nm),  (relative)
-350 2.22e-4
-375 1.40e-3
-400 5.53e-3
-425 7.88e-2 
-450 6.50e-1
-475 7.50e-1
-500 8.73e-1
-525 9.28e-1
-550 9.53e-1
-575 1.00e+0
-600 9.99e-1
-625 9.75e-1
-650 9.38e-1
-675 9.08e-1
-700 8.62e-1
-725 8.03e-1
-750 7.57e-1
-775 6.86e-1
-800 6.55e-1
-825 5.60e-1
-850 4.65e-1
-875 3.90e-1
-900 2.84e-2
-925 5.24e-4
-950 2.07e-4
-
-
