--- python/platosim/varsim/data/response_plato.txt
@@ -1,12 +1,0 @@
-# [nm]	response 	T4000K		T5000K		T6000K		T6500K
-500	0.4848124	0.022262	0.034954	0.045832	0.050504
-550	0.5737143	0.034654	0.047805	0.057579	0.061441
-600	0.6549149	0.048251	0.059793	0.067148	0.069786
-650	0.6559062	0.055751	0.063135	0.066871	0.067986
-700	0.661177	0.062184	0.06523		0.065752	0.065619
-750	0.5891171	0.059388	0.058333	0.056362	0.055366
-800	0.5140042	0.054192	0.050283	0.046841	0.04539
-850	0.3818507	0.041307	0.036468	0.032909	0.031514
-900	0.2484155	0.02716		0.022952	0.020143	0.019091
-950	0.1488914	0.016257	0.013218	0.011318	0.010629
-1000	0.049204	0.005315	0.004175	0.003498	0.003258
