--- python/platosim/varsim/data/response_plato_v2.txt
@@ -1,23 +1,0 @@
-# As Required Values from:
-# Column 1: Wavelenght [nm] 
-# Column 2: Optical transmission of the N-CAM at BOL
-# Column 3: Optical transmission of the N-CAM at EOL
-# Column 4: Optical transmission of the F-CAM blue
-# Column 5: Optical transmission of the F-CAM red
-# Column 6: Quantum efficiency
-500 	0.658	0.634 	nan	nan 	0.730	
-505 	nan 	nan	0.559	nan	nan
-525	nan	nan	0.582	nan	nan
-550	0.711	0.688	0.604	nan	0.800
-575	nan	nan	0.620	nan	nan
-600	0.747	0.727	0.635	nan	0.875
-625	nan	nan	0.639	nan	nan
-650	0.756 	0.738	0.643	nan	0.870
-665	nan	nan	nan	0.647	nan
-700	0.771	0.755	0.655	0.655	0.860
-750	0.779	0.763	nan 	0.662	0.750
-800	0.785	0.770	nan 	0.667	0.650
-850	0.788	0.773	nan	0.670	0.490
-900	0.791	0.776	nan	0.672	0.300
-950	0.792	0.777	nan 	0.673	0.170
-1000	0.794	0.778	nan 	0.675	0.045
